Fix a test that used the wrong YAML manifest
The GatewaySecretReferenceGrantAllInNamespace test used the
GatewaySecretReferenceGrantSpecific test's manifest.

This commit also changes the name of this test's Gateway so it's
distinct from the other tests' Gateways. This helps prevent
eventual-consistency problems with the test suite. If tests share
resources then one test can delete the shared resource, but when the
next test runs the cluster hasn't yet processed the delete and the
test setup code gets confused.