Make test resources unique
The tests can get confused if tests share per-test resources. In this
case one test deleted a Gateway which the next test also used. Because
k8s is eventually-consistent, when the next test ran, the "Get" (that
it uses to decide whether to create or update) succeeded because the
delete from the previous test hadn't finished yet, so the test decided
to do an update, but then the delete finished, so the update failed.