Implement a conformance test timeout flag
Many timeouts were hard-coded to 60s which is a long time. I like to
run the suite single-threaded but my system typically doesn't need
anywhere neat that long to settle so this commit allows you to add a
"-timeout=15" to your command line (after the "-args") to get a 15s
timeout instead.